@@ -15,11 +15,12 @@
 package transferer
 
 import (
+	"errors"
+	"fmt"
 	"io"
 	"os"
 	"testing"
 
-	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
@@ -46,7 +47,7 @@ func (s fakeSource) ReadChunks(m dump.ChunkMeta) ([]*dump.Chunk, error) {
 func (s fakeSource) WriteChunk(_ string, r io.Reader) error {
 	chunkContent, err := io.ReadAll(r)
 	if err != nil {
-		return errors.Wrap(err, "failed to read chunk content")
+		return fmt.Errorf("failed to read chunk content: %w", err)
 	}
 	if len(chunkContent) == 0 {
 		return errors.New("chunk content is empty")
