@@ -18,6 +18,7 @@ import (
 	"bytes"
 	"compress/gzip"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -25,7 +26,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
 	"github.com/valyala/fasthttp"
 
@@ -66,7 +66,7 @@ const requestTimeout = time.Second * 30
 // If splitting the chunks fails, the only way to export is to increase or remove(set it to 0) `search.maxSamplesPerQuery`/.
 func (s Source) splitChunk(m dump.ChunkMeta) ([]*dump.Chunk, error) {
 	if m.End.UnixMilli()-m.Start.UnixMilli() <= 1 {
-		return nil, errors.New("Time range is less than milliseconds, split is impossible, can only be fixed by increasing -search.maxSamplesPerQuery in VM or setting it to 0.")
+		return nil, errors.New("time range is less than milliseconds, split is impossible, can only be fixed by increasing -search.maxSamplesPerQuery in VM or setting it to 0")
 	}
 
 	dur := m.End.Sub(*m.Start) / 2 //nolint:mnd
@@ -88,11 +88,11 @@ func (s Source) splitChunk(m dump.ChunkMeta) ([]*dump.Chunk, error) {
 
 	firstPart, err := s.ReadChunks(firstMeta)
 	if err != nil {
-		return nil, errors.Wrap(err, "Failed to read first chunk. Start time:"+m.Start.String()+"End time: "+t.String())
+		return nil, fmt.Errorf("failed to read first chunk. Start time: %s, End time: %s: %w", m.Start.String(), t.String(), err)
 	}
 	secondPart, err := s.ReadChunks(secondMeta)
 	if err != nil {
-		return nil, errors.Wrap(err, "Failed to read second chunk. Start time:"+t.String()+"End time: "+m.End.String())
+		return nil, fmt.Errorf("failed to read second chunk. Start time: %s, End time: %s: %w", t.String(), m.End.String(), err)
 	}
 	firstPart = append(firstPart, secondPart...)
 	return firstPart, nil
@@ -101,25 +101,25 @@ func (s Source) splitChunk(m dump.ChunkMeta) ([]*dump.Chunk, error) {
 func (s Source) ReadChunks(m dump.ChunkMeta) ([]*dump.Chunk, error) {
 	body, status, err := ReadChunk(s.c, m.Start, m.End, s.cfg.NativeData, s.cfg.ConnectionURL, s.cfg.TimeSeriesSelectors)
 	if err != nil {
-		return nil, errors.Wrap(err, "Error getting responce from Victoria Metrics")
+		return nil, fmt.Errorf("error getting responce from Victoria Metrics: %w", err)
 	}
 	if status != fasthttp.StatusOK {
 		if strings.Contains(gzipDecode(body), "cannot select more than -search.maxSamplesPerQuery") {
 			c, err := s.splitChunk(m)
 			if err != nil {
-				return nil, errors.Wrap(err, "Failed to split VM chunk")
+				return nil, fmt.Errorf("failed to split VM chunk: %w", err)
 			}
 			log.Debug().Msg("VM chunk was split into several parts")
 			return c, nil
 		}
-		return nil, errors.Errorf("invalid status %d", status)
+		return nil, fmt.Errorf("invalid status %d", status)
 	}
 
 	log.Debug().Msg("Got successful response from Victoria Metrics")
 
 	metrics, err := ParseMetrics(bytes.NewReader([]byte(gzipDecode(body))))
 	if err != nil {
-		return nil, errors.Wrap(err, "Failed to parse metrics from body")
+		return nil, fmt.Errorf("failed to parse metrics from body: %w", err)
 	}
 	samples := 0
 	for _, m := range metrics {
@@ -178,7 +178,7 @@ func ReadChunk(c *client.Client, startTime, endTime *time.Time, nativeData bool,
 	resp, err := c.DoWithTimeout(req, requestTimeout)
 	defer fasthttp.ReleaseResponse(resp)
 	if err != nil {
-		return nil, 0, errors.Wrap(err, "failed to send HTTP request to victoria metrics")
+		return nil, 0, fmt.Errorf("failed to send HTTP request to victoria metrics: %w", err)
 	}
 
 	body := copyBytesArr(resp.Body())
@@ -211,13 +211,13 @@ const (
 func decompressChunk(content []byte) ([]Metric, error) {
 	r, err := gzip.NewReader(bytes.NewReader(content))
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to create gzip reader")
+		return nil, fmt.Errorf("failed to create gzip reader: %w", err)
 	}
 	defer r.Close() //nolint:errcheck
 
 	metrics, err := ParseMetrics(r)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to parse chunk content")
+		return nil, fmt.Errorf("failed to parse chunk content: %w", err)
 	}
 	return metrics, nil
 }
@@ -228,14 +228,14 @@ func compressChunk(chunk []Metric) ([]byte, error) {
 	for _, metric := range chunk {
 		metricData, err := json.Marshal(metric)
 		if err != nil {
-			return nil, errors.Wrap(err, "failed to marshal metric")
+			return nil, fmt.Errorf("failed to marshal metric: %w", err)
 		}
 		if _, err := w.Write(metricData); err != nil {
-			return nil, errors.Wrap(err, "failed to write gzip data")
+			return nil, fmt.Errorf("failed to write gzip data: %w", err)
 		}
 	}
 	if err := w.Close(); err != nil {
-		return nil, errors.Wrap(err, "failed to close gzip writer")
+		return nil, fmt.Errorf("failed to close gzip writer: %w", err)
 	}
 	return buf.Bytes(), nil
 }
@@ -243,19 +243,19 @@ func compressChunk(chunk []Metric) ([]byte, error) {
 func (s Source) splitChunkContent(chunkContent []byte, limit int) ([][]byte, error) {
 	metrics, err := decompressChunk(chunkContent)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to parse chunk content")
+		return nil, fmt.Errorf("failed to parse chunk content: %w", err)
 	}
 
 	chunks, err := s.splitMetrics([][]Metric{metrics}, limit)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to split metrics")
+		return nil, fmt.Errorf("failed to split metrics: %w", err)
 	}
 
 	data := make([][]byte, 0, len(chunks))
 	for _, chunk := range chunks {
 		compressedContent, err := compressChunk(chunk)
 		if err != nil {
-			return nil, errors.Wrap(err, "failed to compress chunk content")
+			return nil, fmt.Errorf("failed to compress chunk content: %w", err)
 		}
 		data = append(data, compressedContent)
 	}
@@ -282,7 +282,7 @@ func (s Source) splitMetrics(metricChunks [][]Metric, limit int) ([][]Metric, er
 	for _, chunk := range newMetricChunks {
 		compressedData, err := compressChunk(chunk)
 		if err != nil {
-			return nil, errors.Wrap(err, "failed to compress metrics")
+			return nil, fmt.Errorf("failed to compress metrics: %w", err)
 		}
 		if len(compressedData) > limit {
 			return s.splitMetrics(newMetricChunks, limit)
@@ -297,17 +297,17 @@ func (s Source) WriteChunk(filename string, r io.Reader) error {
 	}
 	chunkContent, err := io.ReadAll(r)
 	if err != nil {
-		return errors.Wrap(err, "failed to read chunk content")
+		return fmt.Errorf("failed to read chunk content: %w", err)
 	}
 
 	if s.cfg.ContentLimit > 0 && len(chunkContent) > s.cfg.ContentLimit {
 		chunks, err := s.splitChunkContent(chunkContent, s.cfg.ContentLimit)
 		if err != nil {
-			return errors.Wrap(err, "failed to split chunk content")
+			return fmt.Errorf("failed to split chunk content: %w", err)
 		}
 		for i, chunk := range chunks {
 			if err := s.sendChunk(chunk); err != nil {
-				return errors.Wrapf(err, "failed to send splitted chunk %s/%d", filename, i+1)
+				return fmt.Errorf("failed to send splitted chunk %s/%d: %w", filename, i+1, err)
 			}
 		}
 
@@ -315,7 +315,7 @@ func (s Source) WriteChunk(filename string, r io.Reader) error {
 	}
 
 	if err := s.sendChunk(chunkContent); err != nil {
-		return errors.Wrapf(err, "failed to send chunk %s", filename)
+		return fmt.Errorf("failed to send chunk %s: %w", filename, err)
 	}
 
 	return nil
@@ -342,14 +342,14 @@ func (s Source) sendChunk(content []byte) error {
 	resp, err := s.c.DoWithTimeout(req, requestTimeout)
 	defer fasthttp.ReleaseResponse(resp)
 	if err != nil {
-		return errors.Wrap(err, "failed to send HTTP request to victoria metrics")
+		return fmt.Errorf("failed to send HTTP request to victoria metrics: %w", err)
 	}
 
 	if s := resp.StatusCode(); s != fasthttp.StatusOK && s != fasthttp.StatusNoContent {
 		if s == http.StatusRequestEntityTooLarge {
 			return errors.New(errRequestEntityTooLarge)
 		}
-		return errors.Errorf("non-OK response from victoria metrics: %d: %s", s, gzipDecode(resp.Body()))
+		return fmt.Errorf("non-OK response from victoria metrics: %d: %s", s, gzipDecode(resp.Body()))
 	}
 
 	log.Debug().Msg("Got successful response from Victoria Metrics")
@@ -357,8 +357,11 @@ func (s Source) sendChunk(content []byte) error {
 }
 
 func ErrIsRequestEntityTooLarge(err error) bool {
-	if err.Error() == errRequestEntityTooLarge || errors.Cause(err).Error() == errRequestEntityTooLarge {
-		return true
+	for err != nil {
+		if err.Error() == errRequestEntityTooLarge {
+			return true
+		}
+		err = errors.Unwrap(err)
 	}
 	return false
 }
@@ -372,11 +375,11 @@ func (s Source) FinalizeWrites() error {
 
 	status, body, err := s.c.GetWithTimeout(url, requestTimeout)
 	if err != nil {
-		return errors.Wrap(err, "failed to send HTTP request to victoria metrics")
+		return fmt.Errorf("failed to send HTTP request to victoria metrics: %w", err)
 	}
 
 	if status != fasthttp.StatusOK {
-		return errors.Errorf("non-OK response from victoria metrics: %d: %s", status, string(body))
+		return fmt.Errorf("non-OK response from victoria metrics: %d: %s", status, string(body))
 	}
 
 	log.Debug().Msg("Got successful response from Victoria Metrics")
@@ -415,7 +418,7 @@ func (s Source) HasMetrics(start, end time.Time) (bool, error) {
 	defer fasthttp.ReleaseResponse(resp)
 
 	if err != nil {
-		return false, errors.Wrap(err, "failed to send HTTP request to victoria metrics")
+		return false, fmt.Errorf("failed to send HTTP request to victoria metrics: %w", err)
 	}
 
 	body := gzipDecode(copyBytesArr(resp.Body()))
@@ -424,13 +427,13 @@ func (s Source) HasMetrics(start, end time.Time) (bool, error) {
 		if strings.Contains(body, "cannot select more than -search.maxSamplesPerQuery") {
 			return s.splitContainsMetrics(start, end, body)
 		}
-		return false, errors.Errorf("non-OK response from victoria metrics: %d: %s", status, body)
+		return false, fmt.Errorf("non-OK response from victoria metrics: %d: %s", status, body)
 	}
 	log.Debug().Msg("Got successful response from Victoria Metrics")
 
 	metricsResp := new(MetricResponse)
 	if err := json.Unmarshal([]byte(body), metricsResp); err != nil {
-		return false, errors.Wrap(err, "failed to unmarshal metrics response")
+		return false, fmt.Errorf("failed to unmarshal metrics response: %w", err)
 	}
 
 	if metricsResp.Stats.SeriesFetched == "0" {
@@ -448,7 +451,7 @@ func (s Source) splitContainsMetrics(start, end time.Time, body string) (bool, e
 	log.Debug().Msg("Time range for first part: Start:" + start.String() + " End:" + middle.String() + " ")
 	firstPart, err := s.HasMetrics(start, middle)
 	if err != nil && !strings.Contains(body, "cannot select more than -search.maxSamplesPerQuery") {
-		return false, errors.Errorf("non-OK response from victoria metrics: %s", body)
+		return false, fmt.Errorf("non-OK response from victoria metrics: %s", body)
 	}
 
 	if firstPart {
@@ -458,13 +461,13 @@ func (s Source) splitContainsMetrics(start, end time.Time, body string) (bool, e
 	log.Debug().Msg("Time range for first part: Start:" + middle.String() + "End:" + end.String() + " ")
 	secondPart, err := s.HasMetrics(middle, end)
 	if err != nil && !strings.Contains(body, "cannot select more than -search.maxSamplesPerQuery") {
-		return false, errors.Errorf("non-OK response from victoria metrics: %s", body)
+		return false, fmt.Errorf("non-OK response from victoria metrics: %s", body)
 	}
 	if secondPart {
 		log.Debug().Msg("Second part has metrics")
 		return true, nil
 	}
-	return false, errors.New("Can't find metrics")
+	return false, errors.New("can't find metrics")
 }
 
 func SplitTimeRangeIntoChunks(start, end time.Time, delta time.Duration) []dump.ChunkMeta {
